+++ source/ProcessManager.Core.Tests/Fixtures/ProcessManagerCoreAppFixture.cs
@@ -1,0 +1,134 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Diagnostics.CodeAnalysis;
+using Energinet.DataHub.Core.DurableFunctionApp.TestCommon.DurableTask;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.Azurite;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ResourceProvider;
+using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures;
+using Microsoft.Azure.WebJobs.Extensions.DurableTask;
+using Xunit.Abstractions;
+
+namespace Energinet.DataHub.ProcessManager.Core.Tests.Fixtures;
+
+/// <summary>
+/// Support testing ProcessManager.Core functionaility by starting the applications
+/// ProcessManager.Example.Orchestrations and ProcessManager (Api).
+/// </summary>
+public class ProcessManagerCoreAppFixture : IAsyncLifetime
+{
+    private const string TaskHubName = "ProcessManagerCoreAppTest01";
+
+    public ProcessManagerCoreAppFixture()
+    {
+        DatabaseManager = new ProcessManagerDatabaseManager("ProcessManagerCoreAppTests");
+        AzuriteManager = new AzuriteManager(useOAuth: true);
+
+        IntegrationTestConfiguration = new IntegrationTestConfiguration();
+
+        ExampleOrchestrationsAppManager = new ExampleOrchestrationsAppManager(
+            DatabaseManager,
+            IntegrationTestConfiguration,
+            AzuriteManager,
+            taskHubName: TaskHubName,
+            appPort: 8401,
+            manageDatabase: false,
+            manageAzurite: false);
+
+        ProcessManagerAppManager = new ProcessManagerAppManager(
+            DatabaseManager,
+            IntegrationTestConfiguration,
+            AzuriteManager,
+            taskHubName: TaskHubName,
+            appPort: 8402,
+            manageDatabase: false,
+            manageAzurite: false);
+
+        DurableTaskManager = new DurableTaskManager(
+            "AzuriteConnectionString",
+            AzuriteManager.FullConnectionString);
+
+        ServiceBusResourceProvider = new ServiceBusResourceProvider(
+            ExampleOrchestrationsAppManager.TestLogger,
+            IntegrationTestConfiguration.ServiceBusFullyQualifiedNamespace,
+            IntegrationTestConfiguration.Credential);
+    }
+
+    public IntegrationTestConfiguration IntegrationTestConfiguration { get; }
+
+    public ExampleOrchestrationsAppManager ExampleOrchestrationsAppManager { get; }
+
+    public ProcessManagerAppManager ProcessManagerAppManager { get; }
+
+    [NotNull]
+    public IDurableClient? DurableClient { get; private set; }
+
+    [NotNull]
+    public TopicResource? ProcessManagerTopic { get; private set; }
+
+    [NotNull]
+    public TopicResource? EdiTopic { get; private set; }
+
+    private ProcessManagerDatabaseManager DatabaseManager { get; }
+
+    private AzuriteManager AzuriteManager { get; }
+
+    private DurableTaskManager DurableTaskManager { get; }
+
+    private ServiceBusResourceProvider ServiceBusResourceProvider { get; }
+
+    public async Task InitializeAsync()
+    {
+        AzuriteManager.CleanupAzuriteStorage();
+        AzuriteManager.StartAzurite();
+
+        await DatabaseManager.CreateDatabaseAsync();
+
+        var processManagerTopicBuilder = ServiceBusResourceProvider.BuildTopic("pm-topic");
+        ExampleOrchestrationsAppManager.ProcessManagerTopicResources.AddSubscriptionsToTopicBuilder(processManagerTopicBuilder);
+        ProcessManagerAppManager.ProcessManagerTopicResources.AddSubscriptionsToTopicBuilder(processManagerTopicBuilder);
+        ProcessManagerTopic = await processManagerTopicBuilder.CreateAsync();
+
+        var ediTopicBuilder = ServiceBusResourceProvider.BuildTopic("edi-topic");
+        EdiTopic = await ediTopicBuilder.CreateAsync();
+
+        await ExampleOrchestrationsAppManager.StartAsync(
+            ExampleOrchestrationsAppManager.ProcessManagerTopicResources.CreateFromTopic(ProcessManagerTopic),
+            ExampleOrchestrationsAppManager.EdiTopicResources.CreateFromTopic(EdiTopic));
+
+        await ProcessManagerAppManager.StartAsync(
+            ProcessManagerAppManager.ProcessManagerTopicResources.CreateFromTopic(ProcessManagerTopic));
+
+        // Create durable client when TaskHub has been created
+        DurableClient = DurableTaskManager.CreateClient(taskHubName: TaskHubName);
+    }
+
+    public async Task DisposeAsync()
+    {
+        await ExampleOrchestrationsAppManager.DisposeAsync();
+        await ProcessManagerAppManager.DisposeAsync();
+        await DurableTaskManager.DisposeAsync();
+        await DatabaseManager.DeleteDatabaseAsync();
+        await ServiceBusResourceProvider.DisposeAsync();
+
+        AzuriteManager.Dispose();
+    }
+
+    public void SetTestOutputHelper(ITestOutputHelper? testOutputHelper)
+    {
+        ExampleOrchestrationsAppManager.SetTestOutputHelper(testOutputHelper);
+        ProcessManagerAppManager.SetTestOutputHelper(testOutputHelper);
+    }
+}