+++ docs/ProcessManager.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,13 @@
 # ProcessManager.Client Release Notes
 
+## Version 0.25.1
+
+- No functional changes.
+
+## Version 0.25.0
+
+- Add `IBusinessValidatedDto` and `ValidationErrorDto` used to support business validation.
+
 ## Version 0.24.0
 
 - Add `INotifyDataDto` type constraint to `NotifyOrchestrationInstanceV1`