+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,8 +1,26 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
-
 ## Version 0.15.3
 
-- No functional changes.
+- Updated `MeteredDataForMeteringPointRejectedV1` 
+
+## Version 0.17.1
+
+- Update `RequestCalculatedEnergyTimeSeriesInputV1` with the required input.
+- Update `RequestCalculatedEnergyTimeSeriesAcceptedV1` with the correct properties.
+
+## Version 0.17.0
+
+- Add shared namespace for `BRS_026`/`BRS_028` types, so they now live in `BRS_026_028.BRS_026` and `BRS_026_028.BRS_028` namespaces.
+
+## Version 0.16.1
+
+- Add `NotifyEnqueueFinishedV1`.
+
+## Version 0.16.0
+
+- Add `SettlementMethod` and `SettlementVersion` DataHub types.
+- Update `RequestCalculatedEnergyTimeSeries` model with `IBusinessValidatedDto` interface.
+- Update `RequestCalculatedEnergyTimeSeriesRejectedV1` properties to be a list of validation errors.
 
 ## Version 0.15.2